changed layout
