+++ public/style.css
@@ -1,3 +1,5 @@
+
+/* main screen */
 * {
     margin: 0;
     padding: 0;
@@ -18,6 +20,8 @@ body {
     background-repeat: no-repeat;
     background-size: cover;
 }
+
+/* animation styles */
 
 @keyframes body {
     0% {
@@ -188,6 +192,7 @@ body {
     border: white;  
   }
   
+/* animation joke */
 
   @keyframes fade-in {
     100% {
@@ -269,6 +274,7 @@ body {
     .city_input {
         margin-top: 8px;
     }
+
     @media (orientation: landscape) {
         body {
             height: auto;
@@ -338,12 +344,13 @@ body {
     }
 }
 
- /* Extra small devices (phones, 767px and down) */
+ /* Small devices (phones, 767px and down) */
 @media only screen and (max-width: 767px) {
 
     body {
         background-size: auto;
         height: auto;
+        width: auto;
     }
     
 
@@ -377,12 +384,11 @@ body {
     }
 }
 
-/* Small devices (portrait tablets and large phones, 600px and up) */
+/* Extra small devices (portrait tablets and large phones, 600px and up) */
 @media only screen and (max-width: 414px) {
    
     body {
         background-size: auto;
-        height: auto;
     }
 
     .user_interface {
@@ -413,7 +419,7 @@ body {
     }
 
     .icon_weather {
-        width: 210px;
+        width: 150px;
     }
 
    .weekday {
@@ -424,6 +430,7 @@ body {
         align-items: center;
         flex-direction: column;
         font-weight: 300;
+        font-size: 14px;
    }
 
    .temperature {
@@ -431,15 +438,19 @@ body {
    }
 
    .weather_bottom {
-       padding: 0;
-       border-radius: 0;
+        padding: 0;
+        border-radius: 0;
+        display: flex;
+        flex-direction: row;
+        align-items: flex-start;
+        padding: 3px;
    }
 
    .loading {
-       align-items: center;
+        margin-top: 6px;
    }
    #date {
-       margin: 8px;
+       margin: 6px;
    }
 
    #joke h3 {
@@ -452,6 +463,11 @@ body {
    
    .weather_top {
        padding: 0;
+   }
+
+   #status {
+    font-size: 1rem;
+    margin-bottom: 6px;
    }
 }
